backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pandas
python-multipart
requests
openai
python-dotenv
aiofiles
regex
